Add Reddit source

+++ app/src/main/java/com/ammar/wallflow/data/repository/WallpapersRemoteMediator.kt
@@ -6,31 +6,61 @@ import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import androidx.room.withTransaction
 import com.ammar.wallflow.data.db.AppDatabase
+import com.ammar.wallflow.data.db.dao.reddit.RedditSearchQueryWallpapersDao
+import com.ammar.wallflow.data.db.dao.reddit.RedditWallpapersDao
+import com.ammar.wallflow.data.db.dao.wallhaven.WallhavenSearchQueryWallpapersDao
+import com.ammar.wallflow.data.db.dao.wallhaven.WallhavenWallpapersDao
+import com.ammar.wallflow.data.db.entity.OnlineSourceWallpaperEntity
+import com.ammar.wallflow.data.db.entity.reddit.RedditSearchQueryWallpaperEntity
+import com.ammar.wallflow.data.db.entity.reddit.RedditWallpaperEntity
 import com.ammar.wallflow.data.db.entity.search.SearchQueryEntity
-import com.ammar.wallflow.data.db.entity.wallhaven.WallhavenSearchQueryRemoteKeyEntity
+import com.ammar.wallflow.data.db.entity.search.SearchQueryRemoteKeyEntity
 import com.ammar.wallflow.data.db.entity.wallhaven.WallhavenSearchQueryWallpaperEntity
 import com.ammar.wallflow.data.db.entity.wallhaven.WallhavenWallpaperEntity
+import com.ammar.wallflow.data.network.OnlineSourceNetworkDataSource
+import com.ammar.wallflow.data.network.RedditNetworkDataSource
 import com.ammar.wallflow.data.network.WallhavenNetworkDataSource
+import com.ammar.wallflow.data.network.model.OnlineSourceWallpapersNetworkResponse
+import com.ammar.wallflow.data.network.model.reddit.NetworkRedditSearchResponse
+import com.ammar.wallflow.data.network.model.reddit.toWallpaperEntities
+import com.ammar.wallflow.data.network.model.wallhaven.NetworkWallhavenWallpapersResponse
 import com.ammar.wallflow.data.network.model.wallhaven.toWallpaperEntity
+import com.ammar.wallflow.json
+import com.ammar.wallflow.model.search.RedditSearch
+import com.ammar.wallflow.model.search.Search
 import com.ammar.wallflow.model.search.WallhavenSearch
 import java.io.IOException
 import kotlinx.datetime.Clock
+import kotlinx.serialization.encodeToString
 import retrofit2.HttpException
 
 @OptIn(ExperimentalPagingApi::class)
-class WallpapersRemoteMediator(
-    private val search: WallhavenSearch,
+class WallpapersRemoteMediator<T : Search, U : OnlineSourceWallpaperEntity>(
+    private val search: T,
     private val appDatabase: AppDatabase,
-    private val wallHavenNetwork: WallhavenNetworkDataSource,
+    private val network: OnlineSourceNetworkDataSource,
     private val clock: Clock = Clock.System,
-) : RemoteMediator<Int, WallhavenWallpaperEntity>() {
-    private val wallpapersDao = appDatabase.wallhavenWallpapersDao()
+) : RemoteMediator<Int, U>() {
+    private val wallpapersDao = when (search) {
+        is WallhavenSearch -> appDatabase.wallhavenWallpapersDao()
+        is RedditSearch -> appDatabase.redditWallpapersDao()
+        else -> throw RuntimeException()
+    }
+    private val searchQueryWallpapersDao = when (search) {
+        is WallhavenSearch -> appDatabase.wallhavenSearchQueryWallpapersDao()
+        is RedditSearch -> appDatabase.redditSearchQueryWallpapersDao()
+        else -> throw RuntimeException()
+    }
     private val searchQueryDao = appDatabase.searchQueryDao()
-    private val remoteKeysDao = appDatabase.wallhavenSearchQueryRemoteKeysDao()
-    private val searchQueryWallpapersDao = appDatabase.wallhavenSearchQueryWallpapersDao()
+    private val remoteKeysDao = appDatabase.searchQueryRemoteKeysDao()
 
     override suspend fun initialize(): InitializeAction {
-        val searchQueryEntity = searchQueryDao.getBySearchQuery(search.toJson())
+        val searchQueryString = when (search) {
+            is WallhavenSearch -> json.encodeToString<WallhavenSearch>(search)
+            is RedditSearch -> json.encodeToString<RedditSearch>(search)
+            else -> throw RuntimeException()
+        }
+        val searchQueryEntity = searchQueryDao.getBySearchQuery(searchQueryString)
         val lastUpdatedOn = searchQueryEntity?.lastUpdatedOn
             ?: return InitializeAction.LAUNCH_INITIAL_REFRESH
         val cacheTimeout = 3 // hours
@@ -48,10 +78,15 @@ class WallpapersRemoteMediator(
 
     override suspend fun load(
         loadType: LoadType,
-        state: PagingState<Int, WallhavenWallpaperEntity>,
+        state: PagingState<Int, U>,
     ): MediatorResult {
         return try {
-            val searchQueryEntity = searchQueryDao.getBySearchQuery(search.toJson())
+            val queryString = when (search) {
+                is WallhavenSearch -> json.encodeToString<WallhavenSearch>(search)
+                is RedditSearch -> json.encodeToString<RedditSearch>(search)
+                else -> throw RuntimeException()
+            }
+            val searchQueryEntity = searchQueryDao.getBySearchQuery(queryString)
             val nextPage = when (loadType) {
                 LoadType.REFRESH -> null
                 LoadType.PREPEND -> return MediatorResult.Success(endOfPaginationReached = true)
@@ -61,25 +96,57 @@ class WallpapersRemoteMediator(
                         ?: return MediatorResult.Success(endOfPaginationReached = true)
                 }
             }
-            val response = wallHavenNetwork.search(search, nextPage?.toIntOrNull())
-            // if at last page, next page is null else current + 1
-            val nextPageNumber = response.meta?.run {
-                if (current_page != last_page) current_page + 1 else null
+            val response = when (network) {
+                is WallhavenNetworkDataSource -> network.search(
+                    search = search as WallhavenSearch,
+                    page = nextPage?.toIntOrNull(),
+                )
+                is RedditNetworkDataSource -> network.search(
+                    search = search as RedditSearch,
+                    after = nextPage,
+                )
+                else -> throw RuntimeException()
+            }
+            val nextPageStr = when (response) {
+                is NetworkWallhavenWallpapersResponse -> {
+                    // if at last page, next page is null else current + 1
+                    response.meta?.run {
+                        if (current_page != last_page) current_page + 1 else null
+                    }
+                }
+                is NetworkRedditSearchResponse -> response.data.after
+                else -> throw RuntimeException()
             }
             appDatabase.withTransaction {
                 val now = clock.now()
                 val searchQueryId = searchQueryEntity?.id ?: searchQueryDao.upsert(
                     SearchQueryEntity(
                         id = 0,
-                        queryString = search.toJson(),
+                        queryString = queryString,
                         lastUpdatedOn = now,
                     ),
                 ).first()
 
                 if (loadType == LoadType.REFRESH) {
                     remoteKeysDao.deleteBySearchQueryId(searchQueryId)
-                    wallpapersDao.deleteAllUniqueToSearchQueryId(searchQueryId)
-                    searchQueryWallpapersDao.deleteBySearchQueryId(searchQueryId)
+                    when (wallpapersDao) {
+                        is WallhavenWallpapersDao -> {
+                            wallpapersDao.deleteAllUniqueToSearchQueryId(searchQueryId)
+                        }
+                        is RedditWallpapersDao -> {
+                            wallpapersDao.deleteAllUniqueToSearchQueryId(searchQueryId)
+                        }
+                        else -> throw RuntimeException()
+                    }
+                    when (searchQueryWallpapersDao) {
+                        is WallhavenSearchQueryWallpapersDao -> {
+                            searchQueryWallpapersDao.deleteBySearchQueryId(searchQueryId)
+                        }
+                        is RedditSearchQueryWallpapersDao -> {
+                            searchQueryWallpapersDao.deleteBySearchQueryId(searchQueryId)
+                        }
+                        else -> throw RuntimeException()
+                    }
                     (searchQueryEntity ?: searchQueryDao.getById(searchQueryId))?.run {
                         // insert or update search query in db
                         searchQueryDao.upsert(copy(lastUpdatedOn = now))
@@ -88,36 +155,87 @@ class WallpapersRemoteMediator(
 
                 // Update RemoteKey for this query.
                 val remoteKey = remoteKeysDao.getBySearchQueryId(searchQueryId)
-                val updatedRemoteKey = remoteKey?.copy(nextPage = nextPageNumber.toString())
-                    ?: WallhavenSearchQueryRemoteKeyEntity(
+                val updatedRemoteKey = remoteKey?.copy(nextPage = nextPageStr.toString())
+                    ?: SearchQueryRemoteKeyEntity(
                         id = 0,
                         searchQueryId = searchQueryId,
-                        nextPage = nextPageNumber.toString(),
+                        nextPage = nextPageStr.toString(),
                     )
                 remoteKeysDao.insertOrReplace(updatedRemoteKey)
 
-                val networkWallpapers = response.data
-                val wallhavenWallpaperIds = networkWallpapers.map { it.id }
-
-                wallpapersDao.insert(networkWallpapers.map { it.toWallpaperEntity() })
-                val wallpaperEntities = wallpapersDao.getByWallhavenIds(wallhavenWallpaperIds)
+                val wallpaperEntities = insertWallpapers(response)
 
                 // update mapping table
-                searchQueryWallpapersDao.insert(
-                    wallpaperEntities.map {
-                        WallhavenSearchQueryWallpaperEntity(
-                            searchQueryId = searchQueryId,
-                            wallpaperId = it.id,
+                when (searchQueryWallpapersDao) {
+                    is WallhavenSearchQueryWallpapersDao -> {
+                        searchQueryWallpapersDao.insert(
+                            @Suppress("UNCHECKED_CAST")
+                            (wallpaperEntities as List<WallhavenWallpaperEntity>).map {
+                                WallhavenSearchQueryWallpaperEntity(
+                                    searchQueryId = searchQueryId,
+                                    wallpaperId = it.id,
+                                )
+                            },
                         )
-                    },
-                )
+                    }
+                    is RedditSearchQueryWallpapersDao -> {
+                        searchQueryWallpapersDao.insert(
+                            @Suppress("UNCHECKED_CAST")
+                            (wallpaperEntities as List<RedditWallpaperEntity>).map {
+                                RedditSearchQueryWallpaperEntity(
+                                    searchQueryId = searchQueryId,
+                                    wallpaperId = it.id,
+                                )
+                            },
+                        )
+                    }
+                    else -> throw RuntimeException()
+                }
             }
 
-            MediatorResult.Success(endOfPaginationReached = nextPageNumber == null)
+            MediatorResult.Success(endOfPaginationReached = nextPageStr == null)
         } catch (e: IOException) {
             MediatorResult.Error(e)
         } catch (e: HttpException) {
             MediatorResult.Error(e)
         }
+    }
+
+    private suspend fun insertWallpapers(
+        response: OnlineSourceWallpapersNetworkResponse,
+    ): List<OnlineSourceWallpaperEntity> {
+        return when (response) {
+            is NetworkWallhavenWallpapersResponse -> insertWallhavenWallpapers(response)
+            is NetworkRedditSearchResponse -> insertRedditWallpapers(response)
+            else -> throw RuntimeException()
+        }
+    }
+
+    private suspend fun insertWallhavenWallpapers(
+        response: NetworkWallhavenWallpapersResponse,
+    ): List<WallhavenWallpaperEntity> {
+        val networkWallpapers = response.data
+        val wallhavenWallpaperIds = networkWallpapers.map { it.id }
+        (wallpapersDao as WallhavenWallpapersDao).insert(
+            networkWallpapers.map { it.toWallpaperEntity() },
+        )
+        return wallpapersDao.getByWallhavenIds(wallhavenWallpaperIds)
+    }
+
+    private suspend fun insertRedditWallpapers(
+        response: NetworkRedditSearchResponse,
+    ): List<RedditWallpaperEntity> {
+        val redditData = response.data
+        val entities = redditData.children
+            .filter {
+                it.data.thumbnail != "default" && !it.data.is_video
+            }
+            .flatMap {
+                // one post can contain multiple wallpapers
+                it.data.toWallpaperEntities()
+            }
+        val redditIds = entities.map { it.redditId }
+        (wallpapersDao as RedditWallpapersDao).insert(entities)
+        return wallpapersDao.getByRedditIds(redditIds)
     }
 }
